chore: rename file holding meta functions
